hotfix: 공통 컴포넌트 모달 오류 해결 (#165)
