+++ src/pages/club-list/ClubListPage.tsx
@@ -15,7 +15,7 @@ import SurveyBox from '@/assets/common/surveyBox.svg?react';
 import SurveyCardArrow from '@/assets/common/surveyCard_arrow.svg?react';
 // import WhoMake from '@/assets/common/whoMake.svg?react';
 import Footer from '@/components/Common/Footer';
-import Modal from '@/components/Common/Modal/Modal';
+import MainModal from '@/components/Common/Modal/MainModal';
 
 // 페이지 컨테이너
 const PageContainer = styled.div`
@@ -496,7 +496,6 @@ const ClubListPage = () => {
                     content: text
                 }
             });
-            console.log('피드백이 성공적으로 제출되었습니다! 🎉');
             // 성공적으로 제출되면 모달을 닫습니다
             setIsSurveyModalOpen(false);
         } catch (error) {
@@ -525,7 +524,7 @@ const ClubListPage = () => {
                         <SurveyBox />
                         <SurveyCardArrow />
                     </SurveyButton>
-                    <Modal 
+                    <MainModal 
                         isOpen={isSurveyModalOpen} 
                         toggle={toggleSurveyModal}
                         title="이용경험을 공유해 주세요."
